chore: fix preexisting ruff lint failures (RUF059/RUF046/RUF043/C420) (#27)
Auto-fixable unused-unpacking, redundant int() casts, raw-string match
patterns and an unsorted import block across analysis, plotting, calibration
and test modules. No behaviour change.

+++ workflow/scripts/plotting/plot_crop_trade_map.py
@@ -632,7 +632,7 @@ def _plot_map(
         top_ref = int(max_mt // 10) * 10
         if top_ref < max_mt * 0.5:
             top_ref = int(max_mt)
-        bot_ref = max(1, int(round(min_mt)))
+        bot_ref = max(1, round(min_mt))
         if bot_ref >= top_ref:
             bot_ref = max(1, top_ref // 4)
